Fix LTO removing SD_EVT_IRQHandler
SD_EVT_IRQHandler is an alias for SWI2_IRQHandler, which is a weak
symbol defined in the gcc_startup ASM code, and positioned in the ISR
vector. The strong SWI2_IRQHandler symobl defined in
app_hello_world/main.c is then meant to override it.

However, with main.c appearing before gcc_startup on the gcc command
line, LTO removes the SD_EVT_IRQHandler as unused. Swapping the order
helps.

See also:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=83967
https://bugs.launchpad.net/gcc-arm-embedded/+bug/1747966
https://bugs.launchpad.net/gcc-arm-embedded/+bug/1788156